move to correct dir
